W: add run-as-admin manifest variant
By default, StartWaitTree is content with the privileges of the executing
user. However, if it is used as part of an installation process and the
child process elevates itself, it becomes impossible for StartWaitTree to
track it (AssignProcessToJobObject fails with ERROR_ACCESS_DENIED). To work
around this, build a variant of StartWaitTree named StartWaitTree_Admin
which differs only in the execution level requested in its manifest
(requireAdministrator, i.e. enforced elevation).